expose location update
