+++ packages/grafana-runtime/src/services/index.ts
@@ -1,3 +1,4 @@
 export * from './backendSrv';
 export * from './AngularLoader';
 export * from './dataSourceSrv';
+export * from './LocationSrv';
