+++ public/app/core/actions/location.ts
@@ -1,4 +1,4 @@
-import { LocationUpdate } from 'app/types';
+import { LocationUpdate } from '@grafana/runtime';
 import { actionCreatorFactory } from 'app/core/redux';
 
 export const updateLocation = actionCreatorFactory<LocationUpdate>('UPDATE_LOCATION').create();
