+++ public/app/routes/GrafanaCtrl.ts
@@ -16,6 +16,9 @@ import { KeybindingSrv, setKeybindingSrv } from 'app/core/services/keybindingSrv
 import { AngularLoader, setAngularLoader } from 'app/core/services/AngularLoader';
 import { configureStore } from 'app/store/configureStore';
 
+import { LocationUpdate, setLocationSrv } from '@grafana/runtime';
+import { updateLocation } from 'app/core/actions';
+
 // Types
 import { KioskUrlValue } from 'app/types';
 
@@ -40,7 +43,12 @@ export class GrafanaCtrl {
     setDataSourceSrv(datasourceSrv);
     setTimeSrv(timeSrv);
     setKeybindingSrv(keybindingSrv);
-    configureStore();
+    const store = configureStore();
+    setLocationSrv({
+      update: (opt: LocationUpdate) => {
+        store.dispatch(updateLocation(opt));
+      },
+    });
 
     $scope.init = () => {
       $scope.contextSrv = contextSrv;
